requirements.txt
gunicorn==19.7.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.0
Flask==2.2.5
Flask-HTTPAuth==4.8.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-streetview==1.2.9
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
instructor==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.2
kwconfig==1.1.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.3
oauthlib==2.1.0
openai==1.6.1
pillow==10.2.0
polyline==2.0.1
proglog==0.1.10
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
requests==2.31.0
requests-oauthlib==1.1.0
rich==13.7.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.2.2
yarl==1.9.4
zipp==3.17.0
opencv-python==4.9.0.80
pytz==2023.3.post1
